The basic gameplay of guessing back and forth now works.
Checking if a ship is totally destroyed still does not work yet. Ending a game
does not work yet.

A bug that allows a player to click on an already hit coordinate and causes
them to lose thier turn....need to fix.

Fixed a bug that caused the ships to be displayed at the wrong xLoc coordinate.