Some visual changes and http redirect.
Https request should now be redirected to http in order for socket.io to work correctly. Before it would only work on http and not https. This still needs to be pushed on Heroku for testing.

The visual changes include some graphics for the landing/enter player name page. The game page itself has gotten some small changes. More visual changes to come.